- renamed convertdata extension to .ts - remove local update variable on historic sim to rely on usegetvehiclebylocation usequery loading state - usesimulatefleet: typing out all exisiting and new any and object types - separate distrubtion graph - and component - separate vehicle selection and input field component - separate simulation header result and kpi metric card component - separate vehicle usage results table and tab
